fix(catalog_service): Fixing the Kafka handle_sneaker event handler
- Added an explicit event type check for calling the delete_sneaker coroutine

+++ backend/catalog_service/catalog_service/catalog/kafka/kafka_handlers/sneaker_handler.py
@@ -22,7 +22,7 @@ async def handle_sneaker(key: str | None, value: dict):
                 sneaker_update = SneakerUpdate(**data)
                 await update_sneaker(session, sneaker_id, sneaker_update)
 
-            else:
+            elif event_type == "sneaker_deleted":
                 sneaker_id = value.get("sneaker_id")
                 await delete_sneaker(session, sneaker_id)
     except Exception as e:
